Direct matches by URL path work again
This commit fixes a recent regression as a result of 8806c9a31488710644975b17a5e558110c265d6f
with the determination of non-pattern vs pattern URLs. That in turn affects the ability to perform
direct matches by URL path.

There is also a fix in PathPattern to recognize "catch-all" patterns as pattern syntax.

See gh-24945